Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)

Funksjon som genererer en figur med andeler av en variabel for en valgt grupperingsvariabel, foreløpig bare sykehus. Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus

figAndelerGrVar(
  RegData = 0,
  hentData = 0,
  preprosess = 0,
  valgtVar = "liggetidPostOp",
  minald = 0,
  maxald = 130,
  erMann = 9,
  datoFra = "2023-01-01",
  datoTil = Sys.Date(),
  Ngrense = 10,
  reshID = 0,
  outfile = "",
  ...
)

Arguments

RegData

dataramme

hentData

hente data? 0-nei, 1-ja

preprosess

preprosessere data? 0-nei, 1-ja

valgtVar

Variabelen det skal vises resultat for.

minald

minimum alder

maxald

maksimum alder

erMann

kjønn 1=mann, 0=kvinne

datoFra

startdato

datoTil

sluttdato

Ngrense

Minste antall registreringer for at ei gruppe skal bli vist

reshID

avdelingas reshid

outfile

Format på figurfila: pdf, png,... ”-standard, viser figur på skjerm

...

mulige tilleggsvariabler

Value

Figur som viser andel av en variabel for gitt grupperingsvariabel. Grupperingsvariabelen (grVar) er som regel den enkelte sykehusenhet.